+++ mirrord/agent/src/incoming/test/http.rs
@@ -1,0 +1,2 @@
+//! Tests for redirecting HTTP requests with the
+//! [`RedirectorTask`](crate::incoming::RedirectorTask).
